+++ tiled/adapters/protocols.py
@@ -4,8 +4,11 @@ from typing import Any, Dict, List, Literal, Optional, Protocol, Set, Tuple, Uni
 
 import dask.dataframe
 import pandas
+import ragged
 import sparse
 from numpy.typing import NDArray
+
+from tiled.structures.ragged import RaggedStructure
 
 from ..ndslice import NDSlice
 from ..storage import Storage
@@ -82,6 +85,22 @@ class AwkwardAdapter(BaseAdapter, Protocol):
         pass
 
 
+class RaggedAdapter(BaseAdapter, Protocol):
+    structure_family: Literal[StructureFamily.ragged]
+
+    @abstractmethod
+    def structure(self) -> RaggedStructure:
+        pass
+
+    @abstractmethod
+    def read(self, slice: NDSlice) -> ragged.array:
+        pass
+
+    @abstractmethod
+    def read_block(self, block: Tuple[int, ...]) -> ragged.array:
+        pass
+
+
 class SparseAdapter(BaseAdapter, Protocol):
     structure_family: Literal[StructureFamily.sparse] = StructureFamily.sparse
 
@@ -127,5 +146,10 @@ class TableAdapter(BaseAdapter, Protocol):
 
 
 AnyAdapter = Union[
-    ArrayAdapter, AwkwardAdapter, ContainerAdapter, SparseAdapter, TableAdapter
+    ArrayAdapter,
+    AwkwardAdapter,
+    ContainerAdapter,
+    RaggedAdapter,
+    SparseAdapter,
+    TableAdapter,
 ]