+++ tiled/structures/core.py
@@ -19,6 +19,7 @@ class StructureFamily(str, enum.Enum):
     array = "array"
     awkward = "awkward"
     container = "container"
+    ragged = "ragged"
     sparse = "sparse"
     table = "table"
 
@@ -64,6 +65,9 @@ STRUCTURE_TYPES = OneShotCachedMap[StructureFamily, type](
         StructureFamily.sparse: lambda: importlib.import_module(
             "...structures.sparse", StructureFamily.__module__
         ).SparseStructure,
+        StructureFamily.ragged: lambda: importlib.import_module(
+            "...structures.ragged", StructureFamily.__module__
+        ).RaggedStructure,
         StructureFamily.container: lambda: importlib.import_module(
             "...structures.container", StructureFamily.__module__
         ).ContainerStructure,